create ModelViewer component and move code from App.tsx into component
